fix(component): implemented observable matcher function for jest, refactored tests
