+++ modules/component/spec/helper/jest.observable-matcher.ts
@@ -1,0 +1,5 @@
+import { defaultAssert, observableMatcher } from './observableMatcher';
+
+export const jestMatcher = observableMatcher(defaultAssert, (a, e) =>
+  expect(a).toStrictEqual(e)
+);
